backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
agno
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0